Update Of Instructions
Changed recommendation for shared GND. Previously it was recommended to connect to power supply, but since the current is low, connecting to the X-min or Y-min endstop is fine and generally easier. There has also been a reported issue when wiring to the powersupply (endstops all stayed triggered), while this may be an isolated incident, there was no reason not to change it.